requirements.txt
google-api-python-client~=2.93.0
apiclient~=1.0.4
google-auth
requests~=2.31.0
pytest~=7.4.0